Move vg mod to a subcommand
